Extracted RepeatOnLifecycle to shared utilities
